+++ GridRL/Content/Ability_Content.cs
@@ -1,0 +1,30 @@
+﻿using System.Windows.Forms;
+using System.Collections.Generic;
+
+namespace GridRL {
+    public class Fireball : Ability {
+        public Fireball() {
+            Name = "fireball";
+            Description = "Big ball of gas";
+            GridHeight = 2;
+            GridWidth = 2;
+        }
+
+        public override bool Use(Creature user) {
+            if(user == Program.player) {
+                Program.waitState = 1;
+                while(Program.waitState == 1) {
+                    Application.DoEvents();
+                }
+                if(Program.waitState == -1) {
+                    return false;
+                }
+                List<int> points = user.DirectionToPoints(Program.lastDirection);
+                FireballEffect e = new FireballEffect(points[0], points[1]);
+                Program.world.Effects.Add(e);
+                Program.lastDirection = Direction.None;
+            }
+            return true;
+        }
+    }
+}
