+++ GridRL/Engine/ImageSprite.cs
