+++ GridRL/GameClasses/Item.cs
@@ -1,5 +1,4 @@
 ﻿using System.Drawing;
-using System.Collections.Generic;
 using System.Linq;
 
 namespace GridRL {
@@ -40,22 +39,6 @@ namespace GridRL {
         public virtual void OnActivate(Creature activator) { }
     }
 
-    public class Weapon : Item {
-        /* Constructors */
-        public Weapon(int y, int x) :base(Properties.Resources.Weap, y, x) {
-            Name = "weapon";
-            Description = "It's some kind of weapon. You can't make out what it is.";
-            MaxStack = 1;
-        }
-
-        /* Properties */
-        public int Attack { get; set; } = 3;
-
-        public float EffectChance { get; set; } = .75f;
-
-        /* Methods */
-        public virtual void OnStrike(Creature owner, Creature struck) { }
-    }
 
     /// <summary> Represents a collection of items for creatures and tiles. </summary>
     public class Inventory {
