+++ GridRL/GridRL.csproj
@@ -46,29 +46,33 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Ability.cs" />
-    <Compile Include="Actor.cs" />
-    <Compile Include="AI.cs" />
-    <Compile Include="Creature.cs" />
-    <Compile Include="DummyCreature.cs" />
-    <Compile Include="Dungeon.cs" />
-    <Compile Include="Engine.cs">
+    <Compile Include="GameClasses\Ability.cs" />
+    <Compile Include="GameClasses\Actor.cs" />
+    <Compile Include="Misc\AI.cs" />
+    <Compile Include="Content\Ability_Content.cs" />
+    <Compile Include="Content\Creature_Content.cs" />
+    <Compile Include="Content\Effect_Content.cs" />
+    <Compile Include="Content\Item_Content.cs" />
+    <Compile Include="Content\Tile_Content.cs" />
+    <Compile Include="GameClasses\Creature.cs" />
+    <Compile Include="Misc\Dungeon.cs" />
+    <Compile Include="Engine\Engine.cs">
       <SubType>Form</SubType>
     </Compile>
-    <Compile Include="Engine.Designer.cs">
+    <Compile Include="Engine\Engine.Designer.cs">
       <DependentUpon>Engine.cs</DependentUpon>
     </Compile>
-    <Compile Include="ImageSprite.cs" />
-    <Compile Include="Item.cs" />
-    <Compile Include="Player.cs" />
-    <Compile Include="Program.cs">
+    <Compile Include="Engine\ImageSprite.cs" />
+    <Compile Include="GameClasses\Item.cs" />
+    <Compile Include="Misc\Player.cs" />
+    <Compile Include="Engine\Program.cs">
       <SubType>Form</SubType>
     </Compile>
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="Sprite.cs" />
-    <Compile Include="Tile.cs" />
-    <Compile Include="World.cs" />
-    <EmbeddedResource Include="Engine.resx">
+    <Compile Include="Engine\Sprite.cs" />
+    <Compile Include="GameClasses\Tile.cs" />
+    <Compile Include="Misc\World.cs" />
+    <EmbeddedResource Include="Engine\Engine.resx">
       <DependentUpon>Engine.cs</DependentUpon>
     </EmbeddedResource>
     <EmbeddedResource Include="Properties\Resources.resx">
@@ -118,6 +122,7 @@
   <ItemGroup>
     <None Include="Resources\Fireball.png" />
   </ItemGroup>
+  <ItemGroup />
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <!-- To modify your build process, add your task inside one of the targets below and uncomment it. 
        Other similar extension points exist, see Microsoft.Common.targets.
